added example for CompactAs and fixed a test

+++ tests/mod.rs
@@ -372,10 +372,10 @@ fn generic_bound_hascompact() {
 	#[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
 	#[derive(PartialEq, Eq, Clone)]
 	// This struct does not impl Codec
-	struct StructHasCompact(u8);
+	struct StructHasCompact(u32);
 
 	impl CompactAs for StructHasCompact {
-		type As = u32;
+		type As = u8;
 		fn encode_as(&self) -> &Self::As {
 			&0
 		}
